object-detection/requirements.txt
opencv-python
Pillow
flask
numpy